shihaja

+++ UserInterface/views.py
@@ -86,7 +86,7 @@ def checking(request):
 	for key, value in dic.items():
 		if value > 0:
 			dt[key] = str(df.dtypes[key])
-	print(dic)
+	print(dt)
 	save_to_config_func(dt, 'missing values datatype', conf)
 	cnt = len(categs)
 	if(num == 0):
@@ -171,20 +171,24 @@ def cleancontinued(request):
 	df = readcsv(pth, 'none')
 
 	#I generated a form that has checkboxes depending on the correlation table
-	if request.method == 'POST' :
+	if request.method == 'POST':
 		categs = getcols(df)
 		form = DropFeat(request.POST)
 		for col in categs:
 			if request.POST.get(col) == 'on':
 				items.append(col)
 				cnt += 1
+		# print
 		if cnt == 0:
-			df.to_csv('fullycleaned.csv', index = False)
-			os.remove("cleaned.csv") 
+			# print("CNT == 0")
+			df.to_csv('fullytransformed.csv', index = False)
+			if os.path.exists("cleaned.csv"):
+				os.remove("cleaned.csv") 
 			message = "No columns will be dropped"
 			return render(request, 'cleancontinued.html', {'form': form, 'message': message, 'cnt': cnt})
 		else:
 			#We make changes on df aka drop the selected columns and return the new df
+			# print("CNT != 0")
 			save_to_config_func(items, 'colstodrop', conf)
 			newdf = dropcols(items, df)
 			#Haven't decided yet whether we should delete cleaned.csv and keep fullycleaned or keep them both
@@ -208,34 +212,38 @@ def correlation(request):
 	message = ""
 	pth = curr + "/UserInterface" 
 	df = readcsv(pth, 'fullycleaned.csv')
+	# print(df)
 
 	#I generated a form that has checkboxes depending on the correlation table
 	# print(df)
 	if request.method == 'POST' :
 		form = TransformForm(request.POST)
 		transform = request.POST.get('technique')
+
+		print(transform)
 		
-		#Here we will be transforming our categorical values
-		obj_df = df.select_dtypes(include=['object']).copy()
-		for col in obj_df:
-			df[col] = df[col].astype('category')
-			pd.get_dummies(df, columns=[col])
+		# #Here we will be transforming our categorical values
+		# obj_df = df.select_dtypes(include=['object']).copy()
+		# for col in obj_df:
+		# 	df[col] = df[col].astype('category')
+		# 	pd.get_dummies(df, columns=[col])
 
-		#TODO separate the normal categories from datetime formats
-		for col in df:
-			if col == "date":
-				x = df[col].str.len()
-				# print(x[2])
-				if x[1] == 15:
-					df[col] = df[col].str.slice(0, 8)
-					df[col] = df[col].astype(int)
-
+		# #TODO separate the normal categories from datetime formats
+		# for col in df:
+		# 	if col == "date":
+		# 		x = df[col].str.len()
+		# 		# print(x[2])
+		# 		if x[1] == 15:
+		# 			df[col] = df[col].str.slice(0, 8)
+		# 			df[col] = df[col].astype(int)
+		
 		if(transform == '1'): #PCA
 			tar = get_data('prediction',conf)
 			final = PC(df, tar)
 			# print(final)
 			final.to_csv('transformed.csv', index = False)
 			cnt = colcnt(final)
+			# print("hjere")
 		#2 is LDA and 3 is FA
 		
 	return render(request, 'correlation.html', {'cnt': cnt})
@@ -305,3 +313,24 @@ class DropFeat(forms.Form):
 	sqft_living15 = forms.BooleanField(required = False, label="sqft_living15 - 0.5854",initial = False)
 	sqft_lot15 = forms.BooleanField(required = False, label="sqft_lot15 - 0.0824",initial = False)
 
+class DropFeat(forms.Form):
+	ids = forms.BooleanField(required = False, label="ids - -0.0169",initial = False)
+	date = forms.BooleanField(required = False, label="date - 0.0030",initial = False)
+	bedrooms = forms.BooleanField(required = False, label="bedrooms - 0.3083",initial = False)
+	bathrooms = forms.BooleanField(required = False, label="bathrooms - 0.5252",initial = False)
+	sqft_living = forms.BooleanField(required = False, label="sqft_living - 0.7021",initial = False)
+	sqft_lot = forms.BooleanField(required = False, label="sqft_lot - 0.0897",initial = False)
+	floors = forms.BooleanField(required = False, label="floors - 0.2568",initial = False)
+	waterfront = forms.BooleanField(required = False, label="waterfront - 0.2664",initial = False)
+	view = forms.BooleanField(required = False, label="view - 0.3974",initial = False)
+	condition = forms.BooleanField(required = False, label="condition - 0.0364",initial = False)
+	grade = forms.BooleanField(required = False, label="grade - 0.6674",initial = False)
+	sqft_above = forms.BooleanField(required = False, label="sqft_above - 0.6056",initial = False)
+	sqft_basement = forms.BooleanField(required = False, label="sqft_basement - 0.3238",initial = False)
+	yr_built = forms.BooleanField(required = False, label="yr_built - 0.0539",initial = False)
+	yr_renovated = forms.BooleanField(required = False, label="yr_renovated - 0.1264",initial = False)
+	zipcode = forms.BooleanField(required = False, label="zipcode - -0.0533",initial = False)
+	lat = forms.BooleanField(required = False, label="lat - 0.3070",initial = False)
+	longi = forms.BooleanField(required = False, label="longi - 0.0216",initial = False)
+	sqft_living15 = forms.BooleanField(required = False, label="sqft_living15 - 0.5854",initial = False)
+	sqft_lot15 = forms.BooleanField(required = False, label="sqft_lot15 - 0.0824",initial = False)
